% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getListOfCheatsheets}
\alias{getListOfCheatsheets}
\title{Download a List of available Cheatsheets}
\usage{
getListOfCheatsheets()
}
\value{
dataframe of available cheatsheets
}
\description{
Download a List of available Cheatsheets
}
\examples{
getListOfCheatsheets()
}
