% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{openCheatsheet}
\alias{openCheatsheet}
\title{Open and download cheatsheet in the local browser}
\usage{
openCheatsheet(entry)
}
\arguments{
\item{entry}{One row from getListOfCheatsheets()}
}
\description{
Open and download cheatsheet in the local browser
}
\examples{
openCheatsheet(getListOfCheatsheets()[0])
}
